Consolidate Part 17 materialization into a shared AliasNameNodeMaterializer
DiagnosticsNodeManager (standard well-known nodes) and
AliasNameNodeManager (application namespaces) each carried their own
descriptor walk, capability-to-optional-child dispatch and
BadUserAccessDenied construction — and the browsable AliasNameType
instance nodes existed only in the diagnostics copy, so a server that
adopted Part 17 via its own AliasNameNodeManager still failed the CTT
§6.2 browse unit while the two walkers drifted (capability handling,
LastChange seeding and the mutation-gate policy already differed).

The walk now exists once, in Opc.Ua.Server.AliasNames:

* AliasNameNodeMaterializer walks a store's descriptor tree and
  materializes categories (with their declared optional children at the
  Part 17 §9 reserved NodeIds where allocated), wires all method
  handlers against an IAliasNameStoreRegistry, and creates the
  browsable AliasNameType instance nodes with AliasFor/HasAlias
  references — carrying over every hardening from the review pass
  (idempotent visited-set walk, namespace/occupancy creation guards,
  resolved-target duplicate detection, dotted-name collision fallback,
  ns0 BrowseName re-homing, remote-target ServerIndex).
* IAliasNameMaterializerHost is the small internal surface both
  managers implement over AsyncCustomNodeManager: predefined-node
  lookup/registration, id minting, root linkage (direct for the
  diagnostics manager, external references for a standalone one) and
  LastChange bookkeeping. Policy differences stay with the hosts —
  the diagnostics path always gates mutations on SecurityAdmin, the
  standalone manager honors RequireSecurityAdminForMutations.
* AliasNameNodeManager now materializes alias instance nodes by
  default (new option MaterializeAliasNodes, on by default; disable
  for the previous category-only behavior). Its BuildCategoryTree/
  WireCategoryHandlers/dispatch helpers are gone; OnStoreChanged
  resolves categories via FindPredefinedNode since every materialized
  category is individually registered. Category descriptors must lie
  in the manager's own namespace.
* The reserved-id table and the materializer log messages moved into
  the AliasNames namespace (AliasNameReservedChildIds,
  AliasNameNodeMaterializerLog under the AliasNameNodeManager event-id
  block); DiagnosticsNodeManager.AliasNames.cs shrinks to the
  standard-node binder, the host implementation and the registry
  LastChange mirroring.

Two new unit tests pin the headline behavior: a standalone manager's
aliases are browsable AliasNameType nodes with the AliasFor reference
and queued inverse reference, and MaterializeAliasNodes=false restores
the category-only tree.
